requirements.txt
certifi==2021.5.30
chardet==4.0.0
boto3==1.18.53
botocore==1.21.53
dj-database-url==0.5.0
django-cloudinary-storage==0.3.0
cloudinary==1.26.0
Django==3.2.4
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-storages==1.11.1
docutils==0.17.1
gunicorn==20.1.0
humanize==3.12.0
idna==2.10
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.5
whitenoise==5.3.0


argon2-cffi==21.1.0
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
boto3==1.18.53
botocore==1.21.53
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.1
cloudinary==1.26.0
colorama==0.4.4
comtypes==1.1.10
DateTime==4.3
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-api==0.1.3
django-bootstrap-modal-forms==2.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-heroku==0.3.1
django-storages==1.11.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
humanize==3.12.0
idna==2.10
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
Js2Py==0.71
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multipart==0.2.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pipwin==0.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
PyAudio
pycparser==2.20
pygame==2.0.1
pygeoip==0.3.2
Pygments==2.10.0
pyjsparser==2.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PyPrind==2.11.3
pyrsistent==0.18.0
pySmartDL==1.3.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-gettext==4.0
python-magic==0.4.24
pyttsx3==2.90
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
reportlab==3.6.2
requests==2.25.1
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.0
transaction==3.0.1
tzlocal==2.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
xhtml2pdf==0.2.5
z3c.xmlhttp==0.5.1
zope.browser==2.3
zope.browserpage==4.4.0
zope.component==5.0.1
zope.configuration==4.4.0
zope.contentprovider==4.2.1
zope.contenttype==4.5.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.exceptions==4.4
zope.hookable==5.1.0
zope.i18n==4.8.0
zope.i18nmessageid==5.0.1
zope.interface==5.4.0
zope.location==4.2
zope.pagetemplate==4.6.0
zope.proxy==4.4.0
zope.publisher==6.0.2
zope.schema==6.2.0
zope.security==5.1.1
zope.tal==4.5
zope.tales==5.1
zope.traversing==4.4.1
zope.viewlet==4.2.1


